[styles] Change style on contrast of page links
Link colors are not meeting contrast requirements for accessibility.
Idea is to change the link color to a darker green to meet requirements
for contrast between the link and the background, as well as,
change
the styling of the link text to differentiate it from surrounding
text. Links in the footer needed to be updated to white as well.

- 'Martel' Google Font added to index.php instead of per article page.
- Add link styles to /bmc/less/_general/_typography.less
- Change link color in /bmc/less/_vars/_colors.less
- Change script variable in /bmc/less/_vars/_fonts.less
- Change links in footer to white /bmc/less/_layout/_template.less
- Remove link styles from images /bmc/js/hub.js

Refs: #190