[feat] Can now override database connection
This is necessary for supergroup components to use their own database
connections. In particular, we can now override the Relational model
default database connection. All we have done is essentially remove
the static attribute from the connection variable and function that
initializes it. Needed to make a fix in __call method as well.