[refactor] Object copyTo efficiency upgrade
Rather than get all objects from the old tag and checking to see
if it doesn't exist in the new tag, just do a query to find all objects
association with the old tag that are not currently associated with the
new tag. This requires only one initial database query, then individual
queries to grab each match. Before, it required a database query for
every object associated with the old tag (which could be a lot!)